Add jobs and modify workshops

+++ Code/views/workshops/index.php
@@ -50,7 +50,9 @@ for ( $i = 0; $i < sizeof($workshops); $i++ ) {
                     <th>DeadLine</th>
                     <th>Details</th>
                     <th>Lecturers</th>
+					<?php if( isAdmin() ): ?>
                     <th></th>
+					<?php endif; ?>
                 </tr>
             </thead>
             <tbody>
@@ -62,11 +64,12 @@ for ( $i = 0; $i < sizeof($workshops); $i++ ) {
                     <td><?= $workshop['deadline'] ?></td>
                     <td><?= $workshop['details'] ?></td>
                     <td><?= $workshop['lecturers'] ?></td>
-                    <td>
-                        <a class="btn btn-sm mb-1 btn-dark" href="show.php">Details</a>
-                        <a class="btn btn-sm mb-1 btn-dark" href="edit.php">Edit</a>
-                        <a class="btn btn-sm mb-1 btn-danger" href="delete.php?id=<?= $workshop['id'] ?>">Delete</a>
-                    </td>
+					<?php if ( isAdmin() ): ?>
+						<td>
+							<a class="btn btn-sm mb-1 btn-dark" href="edit.php?id=<?= $workshop['id'] ?>">Edit</a>
+							<a class="btn btn-sm mb-1 btn-danger" href="delete.php?id=<?= $workshop['id'] ?>">Delete</a>
+						</td>
+					<?php endif; ?>
                 </tr>
 			<?php endforeach; ?>
